requirements.txt
absl-py
edward2
gpflow==2.3.1
h5py
matplotlib
jupyter
scikit-learn
numpy
protobuf
scipy
tensorflow==2.8.0
tensorflow-addons
tensorflow-probability
tqdm
robustness-metrics @ git+https://github.com/google-research/robustness_metrics.git@3e714b3dcce386fa3f473dff8a0f19d7108c18fd